advanze front

+++ server/src/infrastructure/datasources/jobs/jobs.datasource.impl.ts
@@ -1,11 +1,16 @@
 import { prisma } from "../../../database/postgresql-prisma";
 import { JobDatasource } from "../../../domain/datasources/jobs/jobs.datasource";
+import { ApplyJobDto } from "../../../domain/dtos/jobs/apply-job.dto";
 import { CreateJobDto } from "../../../domain/dtos/jobs/create-job.dto";
+import { FavoriteJobDto } from "../../../domain/dtos/jobs/favorite-job.dto";
 import { JobUpdateDto } from "../../../domain/dtos/jobs/update-job.dto";
+import { ApplyJobEntity } from "../../../domain/entities/jobs/apply-job.entity";
+import { FavoriteJobEntity } from "../../../domain/entities/jobs/favorite-job.entity";
 import { JobEntity } from "../../../domain/entities/jobs/jobs.entity";
 import { CustomError } from "../../../domain/errors/custom.error";
 
 export class JobDatasourceImpl implements JobDatasource {
+
     public async createJob(job: CreateJobDto): Promise<JobEntity> {
         try {
             const { userId, companyId, ...rest } = job
@@ -17,8 +22,6 @@ export class JobDatasourceImpl implements JobDatasource {
             if (!area) throw CustomError.notFound('Area not found')
 
             if (userId !== company.userId) throw CustomError.forbidden('You are not allowed to create a job for another company')
-
-            // Si una compañía pertenece a un usuario premium, puede crear más de 5 trabajos, de lo contrario, solo puede crear 5 trabajos
 
             const newJob = await prisma.job.create({
                 data: {
@@ -33,8 +36,6 @@ export class JobDatasourceImpl implements JobDatasource {
                 throw CustomError.forbidden('You can only create 5 jobs , upgrade to premium to create more jobs.')
             }
 
-            console.log({ jobs }) // Ahora debería mostrar todos los trabajos de la compañía, incluyendo el recién creado
-
             return JobEntity.fromObject(newJob)
 
         } catch (error) {
@@ -45,9 +46,14 @@ export class JobDatasourceImpl implements JobDatasource {
             throw CustomError.internalServerError()
         }
     }
-    public async getAllJobs(): Promise<JobEntity[]> {
+    public async getAllJobs(page: number, limit: number): Promise<JobEntity[]> {
+        const limitNumber = Number(limit)
+        const offset = (page - 1) * limit
         try {
-            const jobs = await prisma.job.findMany()
+            const jobs = await prisma.job.findMany({
+                take: limitNumber,
+                skip: offset,
+            })
             return jobs.map(job => JobEntity.fromObject(job))
         } catch (error) {
             if (error instanceof CustomError) {
@@ -73,6 +79,11 @@ export class JobDatasourceImpl implements JobDatasource {
     public async updateJob(user: any, id: string, job: JobUpdateDto) {
         try {
             const { id: idUser } = user
+            const { companyId } = job
+
+            const company = await prisma.company.findUnique({ where: { id: companyId } })
+
+            if (company?.userId !== idUser) throw CustomError.forbidden('You are not allowed to update a job for another company')
 
             const jobToUpdate = await prisma.job.findUnique({ where: { id } })
 
@@ -107,6 +118,10 @@ export class JobDatasourceImpl implements JobDatasource {
 
             if (!jobToDelete) throw CustomError.notFound('Job not found')
 
+            const company = await prisma.company.findUnique({ where: { id: jobToDelete.companyId } })
+
+            if (company?.userId !== idUser) throw CustomError.forbidden('You are not allowed to delete a job for another company')
+
             await prisma.job.delete({ where: { id } })
 
             return 'Job deleted'
@@ -118,5 +133,82 @@ export class JobDatasourceImpl implements JobDatasource {
             throw CustomError.internalServerError()
         }
     }
+    public async addFavoriteJob(favorite: FavoriteJobDto): Promise<FavoriteJobEntity> {
+        try {
+            const { userId, jobId } = favorite
+            const user = await prisma.user.findUnique({ where: { id: userId } })
+            if (!user) throw CustomError.notFound('User not found')
+            const job = await prisma.job.findUnique({ where: { id: jobId } })
+            if (!job) throw CustomError.notFound('Job not found')
 
+            const favoriteJob = await prisma.favoriteJob.create({
+                data: {
+                    userId,
+                    jobId
+                }
+            })
+
+            return FavoriteJobEntity.fromObject(favoriteJob)
+        } catch (error) {
+            if (error instanceof CustomError) {
+                throw error
+            }
+            throw CustomError.internalServerError()
+        }
+    }
+    public async getFavoriteJobs(userId: string): Promise<JobEntity[]> {
+        throw new Error("Method not implemented.");
+    }
+    public async removeFavoriteJob(favorite: FavoriteJobDto): Promise<string> {
+        const { id } = favorite
+        try {
+            const favoriteJob = await prisma.favoriteJob.findUnique({ where: { id } })
+            if (!favoriteJob) throw CustomError.notFound('Favorite job not found')
+            await prisma.favoriteJob.delete({ where: { id } })
+            return 'Favorite job deleted'
+        } catch (error) {
+            if (error instanceof CustomError) {
+                throw error
+            }
+            throw CustomError.internalServerError()
+        }
+    }
+    public async applyJob(apply: ApplyJobDto): Promise<ApplyJobEntity> {
+        try {
+            const { userId, jobId } = apply
+            const user = await prisma.user.findUnique({ where: { id: userId } })
+            if (!user) throw CustomError.notFound('User not found')
+            const job = await prisma.job.findUnique({ where: { id: jobId } })
+            if (!job) throw CustomError.notFound('Job not found')
+
+            const applyJob = await prisma.application.create({
+                data: {
+                    userId,
+                    jobId
+                }
+            })
+
+            return ApplyJobEntity.fromObject(applyJob)
+        } catch (error) {
+            if (error instanceof CustomError) {
+                throw error
+            }
+            throw CustomError.internalServerError()
+        }
+    }
+    public async deleteApplyJob(user: any, id: string): Promise<string> {
+        try {
+
+            const application = await prisma.application.findUnique({ where: { id } })
+            if (!application) throw CustomError.notFound('Application not found')
+            await prisma.application.delete({ where: { id } })
+            return 'Application deleted'
+
+        } catch (error) {
+            if (error instanceof CustomError) {
+                throw error
+            }
+            throw CustomError.internalServerError()
+        }
+    }
 }
